ajout du voter pour tagsAffect et controller

+++ app/src/Controller/LogController.php
@@ -30,9 +30,12 @@ class LogController extends AbstractController
      */
     public function show(Log $log): Response
     {
-        return $this->render('log/show.html.twig', [
-            'log' => $log,
-        ]);
+	if(!$this->isGranted('POST_VIEW',$log)){
+		return $this->render('log/show.html.twig', [
+		    'log' => $log,
+		]);
+	}
+	return $this->redirectToRoute('log_index');
     }
 
     /**
@@ -40,12 +43,13 @@ class LogController extends AbstractController
      */
     public function delete(Request $request, Log $log): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$log->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($log);
-            $entityManager->flush();
-        }
-
+	if(!$this->isGranted('POST_VIEW',$log)){
+		if ($this->isCsrfTokenValid('delete'.$log->getId(), $request->request->get('_token'))) {
+		    $entityManager = $this->getDoctrine()->getManager();
+		    $entityManager->remove($log);
+		    $entityManager->flush();
+		}
+	}
         return $this->redirectToRoute('log_index');
     }
 }
